+++ test/vagrant/config/vm_test.rb
@@ -53,12 +53,4 @@ class ConfigVMTest < Test::Unit::TestCase
       assert @config.proc_stack.include?(proc)
     end
   end
-
-  context "deprecated config" do
-    should "raise an error for provisioner=" do
-      assert_raises(Vagrant::Errors::VagrantError) {
-        @config.provisioner = :chef_solo
-      }
-    end
-  end
 end